+++ src/common/components/AppSidebar.tsx
@@ -13,7 +13,7 @@ import { Bus } from '../icon/Bus'
 import GlobalSeasonFilter from './GlobalSeasonFilter'
 import NavLink from './NavLink'
 import { logoutMutation } from '../query/authQueries'
-import { Button, ButtonSize } from './Button'
+import { Button, ButtonSize } from './buttons/Button'
 import { Text } from '../../util/translate'
 import { useMutationData } from '../../util/useMutationData'
 import { pickGraphqlData } from '../../util/pickGraphqlData'
