+++ src/common/components/PageTitle.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { TabsWrapper } from './Tabs'
-import { Button, ButtonSize, ButtonStyle, StyledButton } from './Button'
+import { Button, ButtonSize, ButtonStyle, StyledButton } from './buttons/Button'
 import { Link } from '@reach/router'
 import { ArrowLeftLong } from '../icon/ArrowLeftLong'
 import { history, pathWithQuery } from '../../util/urlValue'
