+++ src/common/input/Dropdown.tsx
@@ -2,7 +2,7 @@ import React, { CSSProperties, useCallback } from 'react'
 import styled, { css } from 'styled-components/macro'
 import { useSelect } from 'downshift'
 import { text } from '../../util/translate'
-import { Button, ButtonSize } from '../components/Button'
+import { Button, ButtonSize } from '../components/buttons/Button'
 import { ArrowDown } from '../icon/ArrowDown'
 import { observer } from 'mobx-react-lite'
 import { InputLabel } from '../components/form'
