+++ src/common/input/FileUploadInput.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { CrossThick } from '../icon/CrossThick'
 import { CircleCheckmark } from '../icon/CircleCheckmark'
-import { Button, ButtonStyle } from '../components/Button'
+import { Button, ButtonStyle } from '../components/buttons/Button'
 import { FlexRow } from '../components/common'
 import Loading from '../components/Loading'
 import { useDropzone } from 'react-dropzone'
