+++ src/common/input/ItemForm.tsx
@@ -3,7 +3,7 @@ import styled, { CSSProperties } from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { get, uniqueId } from 'lodash'
 import { TextInput } from './Input'
-import { Button, ButtonStyle } from '../components/Button'
+import { Button, ButtonStyle } from '../components/buttons/Button'
 import { useOrderedValues } from '../../util/useOrderedValues'
 import { usePromptUnsavedChanges } from '../../util/promptUnsavedChanges'
 import UserHint from '../components/UserHint'
