+++ src/common/table/TableFiltersControl.tsx
@@ -5,7 +5,7 @@ import Input from '../input/Input'
 import { omit } from 'lodash'
 import { ControlGroup } from '../components/form'
 import Dropdown from '../input/Dropdown'
-import { Button, ButtonSize, ButtonStyle, RemoveButton } from '../components/Button'
+import { Button, ButtonSize, ButtonStyle, RemoveButton } from '../components/buttons/Button'
 import { text, Text } from '../../util/translate'
 import { FlexRow } from '../components/common'
 import { SubHeading } from '../components/Typography'
