+++ src/contract/ContractEditor.tsx
@@ -25,7 +25,7 @@ import { get, isEqual, orderBy } from 'lodash'
 import { pickGraphqlData } from '../util/pickGraphqlData'
 import ContractUsers from './ContractUsers'
 import { useContractPage } from './contractUtils'
-import { Button, ButtonSize, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle } from '../common/components/buttons/Button'
 import { FlexRow } from '../common/components/common'
 import FileUploadInput from '../common/input/FileUploadInput'
 import { useUploader } from '../util/useUploader'
