+++ src/departureBlock/DepartureBlockGroupItem.tsx
@@ -6,7 +6,7 @@ import UploadFile from '../common/input/UploadFile'
 import { LoadingDisplay } from '../common/components/Loading'
 import styled from 'styled-components/macro'
 import { DayTypeGroup, getEnabledDayTypes } from './departureBlocksCommon'
-import { Button, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonStyle } from '../common/components/buttons/Button'
 import { useMutationData } from '../util/useMutationData'
 import { removeDepartureBlocks, uploadDepartureBlocksMutation } from './blockDeparturesQuery'
 import { InspectionContext } from '../inspection/InspectionContext'
