+++ src/departureBlock/DepartureBlocks.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useContext, useEffect, useMemo } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { difference } from 'lodash'
-import { Button } from '../common/components/Button'
+import { Button } from '../common/components/buttons/Button'
 import { defaultDayTypeGroup, useDayTypeGroups } from './departureBlocksCommon'
 import DepartureBlockGroupItem from './DepartureBlockGroupItem'
 import { useQueryData } from '../util/useQueryData'
