+++ src/equipmentCatalogue/EditEquipmentCatalogue.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useRef, useState } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import ItemForm from '../common/input/ItemForm'
-import { Button, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonStyle } from '../common/components/buttons/Button'
 import { EquipmentCatalogue, EquipmentCatalogueInput, ProcurementUnit } from '../schema-types'
 import { isEqual } from 'lodash'
 import EquipmentCatalogueFormInput from './EquipmentCatalogueFormInput'
