+++ src/inspection/InspectionActions.tsx
@@ -2,7 +2,7 @@ import React, { CSSProperties, useCallback, useState } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { Inspection, InspectionStatus, InspectionType, Season } from '../schema-types'
-import { Button, ButtonSize, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle } from '../common/components/buttons/Button'
 import {
   useEditInspection,
   useInspectionReports,
