+++ src/inspection/InspectionApprovalSubmit.tsx
@@ -4,7 +4,7 @@ import SelectDate from '../common/input/SelectDate'
 import { ControlGroup, FormColumn, InputLabel } from '../common/components/form'
 import { Inspection } from '../schema-types'
 import { FlexRow } from '../common/components/common'
-import { Button, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonStyle } from '../common/components/buttons/Button'
 import { ActionsWrapper } from '../common/input/ItemForm'
 import styled from 'styled-components/macro'
 import { addDays, max, parseISO } from 'date-fns'
