+++ src/inspection/InspectionConfig.tsx
@@ -6,7 +6,7 @@ import { ControlGroup, FormColumn, InputLabel } from '../common/components/form'
 import { Inspection, InspectionInput, InspectionStatus } from '../schema-types'
 import { MessageContainer, MessageView } from '../common/components/Messages'
 import { FlexRow, PageSection } from '../common/components/common'
-import { Button, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonStyle } from '../common/components/buttons/Button'
 import { ActionsWrapper } from '../common/input/ItemForm'
 import styled from 'styled-components/macro'
 import InspectionTimeline from './InspectionTimeline'
