+++ src/inspection/inspectionDate/InsectionDateList.tsx
@@ -3,7 +3,7 @@ import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { InspectionDate } from '../../schema-types'
 import { Text } from '../../util/translate'
-import { RemoveButton } from '../../common/components/Button'
+import { RemoveButton } from '../../common/components/buttons/Button'
 import { LoadingDisplay } from '../../common/components/Loading'
 import { useMutationData } from '../../util/useMutationData'
 import { removeInspectionDateMutation } from './inspectionDateQuery'
