+++ src/page/AuthGate.tsx
@@ -6,7 +6,7 @@ import { LoadingDisplay } from '../common/components/Loading'
 import { HSLLogoNoText } from '../common/icon/HSLLogoNoText'
 import { Login } from '../common/icon/Login'
 import { Text } from '../util/translate'
-import { ButtonSize, ButtonStyle, StyledButton } from '../common/components/Button'
+import { ButtonSize, ButtonStyle, StyledButton } from '../common/components/buttons/Button'
 import { AUTH_SCOPE, AUTH_URI, CLIENT_ID, REDIRECT_URI } from '../constants'
 import { getAppRoot, history } from '../util/urlValue'
 import ErrorMessages from '../common/components/ErrorMessages'
