+++ src/page/ContractPage.tsx
@@ -7,7 +7,7 @@ import { useQueryData } from '../util/useQueryData'
 import { Text } from '../util/translate'
 import { contractsQuery } from '../contract/contractQueries'
 import { MessageContainer, MessageView } from '../common/components/Messages'
-import { Button, ButtonSize, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle } from '../common/components/buttons/Button'
 import { useStateValue } from '../state/useAppState'
 import { useHasAdminAccessRights } from '../util/userRoles'
 import { ArrowRight } from '../common/icon/ArrowRight'
