+++ src/page/InspectionDatePage.tsx
@@ -5,7 +5,7 @@ import { PageTitle } from '../common/components/PageTitle'
 import { Page, PageSection, PageContainer } from '../common/components/common'
 import { RouteComponentProps } from '@reach/router'
 import InspectionDateForm from '../inspection/inspectionDate/InspectionDateForm'
-import { Button } from '../common/components/Button'
+import { Button } from '../common/components/buttons/Button'
 import InspectionDateList from '../inspection/inspectionDate/InsectionDateList'
 import { useRefetch } from '../util/useRefetch'
 import { useQueryData } from '../util/useQueryData'
