+++ src/page/InspectionsPage.tsx
@@ -4,7 +4,7 @@ import { RouteComponentProps } from '@reach/router'
 import { Page, PageContainer } from '../common/components/common'
 import InspectionsList from '../inspection/InspectionsList'
 import { Plus } from '../common/icon/Plus'
-import { Button } from '../common/components/Button'
+import { Button } from '../common/components/buttons/Button'
 import { navigateWithQueryString } from '../util/urlValue'
 import { MessageContainer, MessageView } from '../common/components/Messages'
 import { getInspectionTypeStrings, useFetchInspections } from '../inspection/inspectionUtils'
