+++ src/postInspection/LoadInspectionHfpData.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useContext, useEffect, useMemo, useState } from 're
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import { PageSection } from '../common/components/common'
-import { Button, ButtonSize } from '../common/components/Button'
+import { Button, ButtonSize } from '../common/components/buttons/Button'
 import { InspectionContext } from '../inspection/InspectionContext'
 import { InputLabel } from '../common/components/form'
 import { gql, useSubscription } from '@apollo/client'
