+++ src/procurementUnit/ProcurementUnits.tsx
@@ -2,7 +2,7 @@ import React, { useCallback, useContext, useState } from 'react'
 import styled from 'styled-components/macro'
 import { observer } from 'mobx-react-lite'
 import ProcurementUnitItem from './ProcurementUnitItem'
-import { Button, ButtonSize, ButtonStyle, TextButton } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle, TextButton } from '../common/components/buttons/Button'
 import { FlexRow, TransparentPageSection } from '../common/components/common'
 import { useQueryData } from '../util/useQueryData'
 import { procurementUnitsQuery } from './procurementUnitsQuery'
