+++ src/report/DownloadReport.tsx
@@ -1,6 +1,6 @@
 import React, { CSSProperties, useCallback, useState } from 'react'
 import { observer } from 'mobx-react-lite'
-import { Button, ButtonSize, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle } from '../common/components/buttons/Button'
 import { SERVER_URL } from '../constants'
 import { saveAs } from 'file-saver'
 import styled from 'styled-components/macro'
