+++ src/report/ReportContainer.tsx
@@ -6,7 +6,7 @@ import { createReportQueryByName } from './reportQueries'
 import { InspectionType } from '../schema-types'
 import { createResponseId, useTableState } from '../common/table/useTableState'
 import DownloadReport from './DownloadReport'
-import { Button, ButtonSize, ButtonStyle } from '../common/components/Button'
+import { Button, ButtonSize, ButtonStyle } from '../common/components/buttons/Button'
 import { Text } from '../util/translate'
 import { FlexRow } from '../common/components/common'
 import { BaseReport } from '../type/report'
